Rename messageTypes.ts -> typing.ts
