+++ frontend/src/networking/typing.ts
